requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
node==1.2.2
odict==1.9.0
orjson==3.10.7
pillow==10.4.0
plumber==1.7
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pymongo==3.12.1
python-dotenv==1.0.1
python-multipart==0.0.10
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.3.1
starlette==0.38.6
typer==0.12.5
typing_extensions==4.12.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==7.0
zope.interface==7.0.3
zope.lifecycleevent==5.0
zope.proxy==6.0
